Lexer: Handle the Potential for Identifiers that Start with a Reserved Keyword
It is possible for identifiers to start with a reserved keyword, for
example "writefoo", previously, this would be lexed as the new keyword
with the remainder being left to be lexed as the identifier foo.
We now correctly don't lex this as a write token. Note that lexing
identifiers is not implemented by this commit, so the example of "writefoo"
would still cause a todo panic.