Compiler: Add Flag to Read and Execute an M File
In order to implement if and else, we will need to be able to support
mult-line programs. Our parser and code gen already did, but our
compiler had no way to feed a file into them. Now we can supply a
file name as a parameter to the program and it will feed that into
the compiler.